requirements.txt
gradio>=4.0.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0
gdown>=4.7.1
